+++ libs/cardano-protocol-tpraos/test/Main.hs
@@ -14,6 +14,7 @@ import Test.Cardano.Ledger.Alonzo.Binary.Annotator ()
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Mary.Binary.Annotator ()
 import Test.Cardano.Ledger.Shelley.Binary.Annotator ()
+import Test.Cardano.Protocol.Binary.Annotator ()
 import qualified Test.Cardano.Protocol.Binary.BinarySpec as Binary
 import qualified Test.Cardano.Protocol.Binary.CddlSpec as Cddl
 import Test.Cardano.Protocol.Binary.RoundTrip
@@ -21,12 +22,11 @@ import Test.Cardano.Protocol.TPraos.Arbitrary ()
 
 main :: IO ()
 main =
-  ledgerTestMain $
-    describe "TPraos" $ do
-      Cddl.spec
-      Binary.spec
-      describe "RoundTrip" $ do
-        roundTripBlockSpec @(BHeader StandardCrypto) @ShelleyEra
-        roundTripBlockSpec @(BHeader StandardCrypto) @AllegraEra
-        roundTripBlockSpec @(BHeader StandardCrypto) @MaryEra
-        roundTripBlockSpec @(BHeader StandardCrypto) @AlonzoEra
+  ledgerTestMain $ do
+    Cddl.spec
+    Binary.spec
+    describe "RoundTrip" $ do
+      roundTripBlockSpec @(BHeader StandardCrypto) @ShelleyEra
+      roundTripBlockSpec @(BHeader StandardCrypto) @AllegraEra
+      roundTripBlockSpec @(BHeader StandardCrypto) @MaryEra
+      roundTripBlockSpec @(BHeader StandardCrypto) @AlonzoEra